0.10 dev Modify synchronization logic

+++ config/config.go
@@ -23,7 +23,7 @@ import (
 
 const (
 	defaultConfigFilename = "config.toml"
-	sampleConfigFilename = "sample-config.toml"
+	sampleConfigFilename  = "sample-config.toml"
 	defaultLogLevel       = "info"
 	defaultLogDirname     = "logs"
 	defaultRPCMaxClients  = 10
@@ -35,7 +35,7 @@ const (
 var (
 	defaultAppDataDir  = utils.AppDataDir("qitwallet", false)
 	DefaultConfigFile  = filepath.Join("./", defaultConfigFilename)
-	SampleConfigFile  = filepath.Join("./", sampleConfigFilename)
+	SampleConfigFile   = filepath.Join("./", sampleConfigFilename)
 	defaultRPCKeyFile  = filepath.Join(defaultAppDataDir, "rpc.key")
 	defaultRPCCertFile = filepath.Join(defaultAppDataDir, "rpc.cert")
 	defaultLogDir      = filepath.Join(defaultAppDataDir, defaultLogDirname)
@@ -66,7 +66,7 @@ type Config struct {
 	DisableRPC    bool
 	DisableTLS    bool
 
-	Confirmations int64
+	Confirmations uint32
 
 	// tx fee
 	MinTxFee int64
@@ -137,7 +137,7 @@ func NewDefaultConfig() (cfg *Config) {
 		AppDataDir: defaultAppDataDir,
 		DebugLevel: defaultLogLevel,
 		LogDir:     defaultLogDir,
-		ConfigFile:"config.toml",
+		ConfigFile: "config.toml",
 
 		Network: "testnet",
 
@@ -162,9 +162,9 @@ func NewDefaultConfig() (cfg *Config) {
 		QProxyUser:     "",
 		QProxyPass:     "",
 		WalletPass:     "public",
-		MinTxFee:		DefaultMinRelayTxFee,
-		Confirmations:		10,
-		UI:true,
+		MinTxFee:       DefaultMinRelayTxFee,
+		Confirmations:  10,
+		UI:             true,
 	}
 	return
 }
