+++ json/block.go
@@ -7,11 +7,11 @@ import (
 
 type BlockHttpResult struct {
 	Hash          string `json:"hash"`
-	Confirmations int64  `json:"confirmations"`
+	Confirmations uint32 `json:"confirmations"`
 	Version       int32  `json:"version"`
 	//Height		  int32			`json:"height"`
 	Weight       int64              `json:"weight"`
-	Order        int32              `json:"order"`
+	Order        uint32             `json:"order"`
 	TxRoot       string             `json:"txRoot"`
 	Transactions []json.TxRawResult `json:"transactions,omitempty"`
 	StateRoot    string             `json:"stateRoot"`
